Fix overwriting variables when running multiple instances.
