--- src/utils/log.js
@@ -1,24 +1,0 @@
-let logging = false;
-let customLogger = () => {};
-
-const setLogging = (_logging) => {
-  logging = _logging;
-};
-
-const setCustomLogger = (logger) => {
-  customLogger = logger;
-};
-
-const log = (type, message) => {
-  customLogger({ type, message });
-  if (logging) {
-    console.log(`[${type}] ${message}`);
-  }
-};
-
-module.exports = {
-  logging,
-  setLogging,
-  setCustomLogger,
-  log,
-};
